break out tests.js into parts.
