+++ tests/legacy.js
@@ -1,0 +1,1 @@
+// These don't run, but they have logic we could use.
